refactor: move fund_address and mine functions to tests-common
