+++ protocol/src/nigiri.rs
@@ -3,7 +3,8 @@
 use crate::protocol_musig_adaptor::MemWallet;
 use bdk_wallet::bitcoin::Amount;
 use std::process::Output;
-use std::{process::Command, thread, time};
+use std::{thread, time};
+use tests_common::{fund_address, mine};
 
 pub(crate) fn funded_wallet() -> MemWallet {
     println!("loading wallet...");
@@ -30,27 +31,8 @@ pub(crate) fn fund_wallet(wallet: &mut MemWallet) {
     assert!(wallet.balance() >= Amount::from_btc(1.0).unwrap());
 }
 
-fn fund_address(address: &str) {
-    let faucet_response = Command::new("nigiri")
-        .args(["faucet", address])
-        .output()
-        .expect("Failed to fund Alice's wallet");
-    eprintln!("{}", String::from_utf8_lossy(&faucet_response.stdout));
-    // thread::sleep(time::Duration::from_secs(2)); // Add delays between steps
-
-    eprintln!("Mining mining to {}", address);
-    let resp = mine(address, 1);
-    eprintln!("reponse {}", String::from_utf8_lossy(&resp.stdout));
-}
-
 const FUND_ADDRESS: &str = "bcrt1plrmcqc9pwf4zjcej5n7ynre5k8lkn0xcz0c7y3dw37e8nqew2utq5l06jv";
 
 pub(crate) fn tiktok() -> Output {
     mine(FUND_ADDRESS, 1)
-}
-fn mine(address: &str, num_blocks: u16) -> Output {
-    Command::new("nigiri")
-        .args(["rpc", "generatetoaddress", &num_blocks.to_string(), address])
-        .output()
-        .expect("Failed to mine block")
 }
